Restructure how IPC is handled
Main modification here is, no work is done in signal handlers. Instead, I'm just setting a global flag for SIGPIPE and SIGCHLD, and polling SIGUSR1 from a new thread that handles outgoing packets to clients.
This has reduced global variables to just 2 atomic flags and an int with the ID of each process, which is needed (by now) for self identification in asynchronous cleanup handlers. Everything else is sent through arguments as structure pointers to shared memory.